Move get_state_at() to area we can share from
